Moved xml_to_csv() to slide.utils.
